ZO: Add character image to LocationAutocomplete
