+++ libs/zzz/ui/src/Character/index.tsx
@@ -1,3 +1,3 @@
-export * from './CharIconSideElement'
+export * from './CharIconCircleElement'
 export * from './LocationAutocomplete'
 export * from './StatDisplay'
